feat: Package-Structure-Changed-checkin
자바 룰을 따르도록 패키지 구조 변경 (Commit 1)
