--- check-in-request-MS/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'check-in-request-MS'
