--- check-in-request-MS/src/main/java/com/example/checkinrequestMS/HelpAPI/domain/model/helper/Helper.java
@@ -1,6 +1,0 @@
-package com.example.checkinrequestMS.HelpAPI.domain.model.helper;
-
-public class Helper {
-    private Long helperId;
-
-}
