--- check-in-request-MS/src/main/java/com/example/checkinrequestMS/HelpAPI/infra/db/help/CheckInSpringJPARepository.java
@@ -1,8 +1,0 @@
-package com.example.checkinrequestMS.HelpAPI.infra.db.help;
-
-import com.example.checkinrequestMS.HelpAPI.infra.db.entity.child.CheckInEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CheckInSpringJPARepository extends JpaRepository<CheckInEntity, Long> {
-
-}
