--- check-in-request-MS/src/main/java/com/example/checkinrequestMS/HelpAPI/infra/db/help/EtcSpringJPARepository.java
@@ -1,9 +1,0 @@
-package com.example.checkinrequestMS.HelpAPI.infra.db.help;
-
-import com.example.checkinrequestMS.HelpAPI.infra.db.entity.child.CheckInEntity;
-import com.example.checkinrequestMS.HelpAPI.infra.db.entity.child.EtcEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface EtcSpringJPARepository extends JpaRepository<EtcEntity, Long> {
-
-}
