--- check-in-request-MS/src/main/java/com/example/checkinrequestMS/PlaceAPI/domain/Type.java
@@ -1,7 +1,0 @@
-package com.example.checkinrequestMS.PlaceAPI.domain;
-
-public enum Type {
-
-    API, DB;
-    
-}
