--- check-in-request-MS/src/main/java/com/example/checkinrequestMS/PlaceAPI/web/restAPI/SearchType.java
@@ -1,13 +1,0 @@
-package com.example.checkinrequestMS.PlaceAPI.web.restAPI;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-
-public enum SearchType {
-    KEYWORD,
-    CATEGORY;
-
-
-}
