--- check-in-request-MS/src/test/java/com/example/checkinrequestMS/Test.java
@@ -1,9 +1,0 @@
-//package com.example.checkinrequestMS;
-//
-//import java.time.LocalDateTime;
-//
-//public class Test {
-//    void test() {
-//
-//    }
-//}
